fix(external-ingestion): PR #23 review blockers — JSONB, SQS publisher, profiles, level_rank
- Apply @JdbcTypeCode(SqlTypes.JSON) to rawCategoryTokens/rawLevelTokens for safe JSONB binding
- Add DisasterAlertJsonbTest (Testcontainers PostgreSQL, skipped without Docker)
- Add SqsCacheEventPublisher: real SQS SendMessage, success/failure metrics, startup fail on missing queueUrl
- Add SqsConfig: SqsClient bean with optional LocalStack endpoint-override
- Switch LoggingCacheEventPublisher to @ConditionalOnProperty(havingValue="false") — startup guard moved from publish-time to boot-time
- Add SqsCacheEventPublisherLocalStackTest (Testcontainers LocalStack, skipped without Docker)
- Add application-dev.yml and application-prod.yml: flyway.enabled=true, ddl-auto=none
- Add SafetyDataAlertNormalizerTest assertion for valid JSON array tokens
- Fix V1 DDL: level_rank INT → SMALLINT

+++ services/external-ingestion/src/main/resources/db/migration/V1__create_initial_schema.sql
@@ -97,7 +97,7 @@ CREATE TABLE disaster_alert (
     -- JSONB: JSON arrays stored as jsonb for indexing and validation
     raw_level_tokens     JSONB,
     level                VARCHAR(10),
-    level_rank           INT,
+    level_rank           SMALLINT,
     region               VARCHAR(100) NOT NULL,
     source_region        VARCHAR(100),
     message              TEXT         NOT NULL,
